BUG-OBS-08: fix monthly-cost query (cast SUM(BIGINT) to bigint)
cost_micros is BIGINT, so Postgres SUM() returns NUMERIC; sqlx 0.8 (built
without bigdecimal/rust_decimal) decodes NUMERIC as String, so the
query_scalar::<Option<i64>> decode failed on every call — masked by .context,
flooding the metrics-collector WARN log and breaking the monthly-spend gauge
and the admin dashboard cost figure (campaign finding F6).

Cast COALESCE(SUM(cost_micros),0)::bigint in both PgStore::get_monthly_cost
(budget.rs) and admin_monthly_spend (admin_queries.rs). Added an #[ignore] pg
regression test (dashboard_pg.rs::monthly_cost_sums_in_month_non_null_micros)
asserting the in-month non-NULL sum, NULL-cost + prior-month exclusion, and the
empty -> 0 COALESCE path. Ledger: budget.rs added to the pg_store conflict group.

+++ crates/store/src/admin_queries.rs
@@ -349,8 +349,12 @@ impl PgStore {
     /// Returns an error if the database is not connected or the query fails.
     pub async fn admin_monthly_spend(&self, tenant_id: i64) -> anyhow::Result<u64> {
         let pool = self.pool()?;
+        // `cost_micros` is BIGINT, so `SUM()` returns NUMERIC, which sqlx 0.8
+        // (no `bigdecimal`/`rust_decimal` feature) cannot decode as `i64`. Cast
+        // back to `bigint` so the `Option<i64>` decode succeeds (mirrors
+        // PgStore::get_monthly_cost).
         let total: Option<i64> = sqlx::query_scalar(
-            "SELECT COALESCE(SUM(cost_micros), 0) \
+            "SELECT COALESCE(SUM(cost_micros), 0)::bigint \
              FROM usage_events \
              WHERE tenant_id = $1 \
                AND cost_micros IS NOT NULL \
